+++ changelog.md
@@ -5,6 +5,7 @@
 * Highlight the row of a selected cell.
 * Fixed error when there is no view.
 * Added Ubuntu installation guide documentation.
+* Added Email field.
 
 ## Released (2020-10-06)